separate out utility functions
